# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chartink Scanner Monitor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Scanner {
  id: number;
  name: string;
  chartinkUrl: string;
  intervalMinutes: number;
  isActive: boolean;
  createdAt: string;
  /** @nullable */
  lastScannedAt: string | null;
  alertCount: number;
  /** @nullable */
  description?: string | null;
}

export interface ScannerInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  chartinkUrl: string;
  intervalMinutes?: number;
  description?: string;
  isActive?: boolean;
}

export interface ScannerUpdate {
  name?: string;
  chartinkUrl?: string;
  intervalMinutes?: number;
  /** @nullable */
  description?: string | null;
  isActive?: boolean;
}

export interface ScannerToggle {
  isActive: boolean;
}

export interface Alert {
  id: number;
  scannerId: number;
  scannerName: string;
  symbol: string;
  /** @nullable */
  price?: number | null;
  triggeredAt: string;
  telegramSent?: boolean;
}

export interface ScanResult {
  stocksFound: number;
  newAlerts: number;
  scannerName: string;
}

export interface StatsSummary {
  totalScanners: number;
  activeScanners: number;
  totalAlerts: number;
  alertsToday: number;
  /** @nullable */
  lastScanAt: string | null;
  telegramEnabled?: boolean;
}

export interface ScannerActivity {
  scannerId: number;
  scannerName: string;
  alertCount: number;
}

export type ListAlertsParams = {
  scannerId?: number;
  symbol?: string;
  limit?: number;
};

export type GetRecentAlertsParams = {
  limit?: number;
};