# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Chartink Scanner Monitor API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: scanners
    description: Scanner management
  - name: alerts
    description: Alert history
  - name: stats
    description: Dashboard statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /scanners:
    get:
      operationId: listScanners
      tags: [scanners]
      summary: List all scanners
      responses:
        "200":
          description: List of scanners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scanner"
    post:
      operationId: createScanner
      tags: [scanners]
      summary: Add a new scanner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScannerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scanner"

  /scanners/{id}:
    get:
      operationId: getScanner
      tags: [scanners]
      summary: Get scanner by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scanner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scanner"
        "404":
          description: Not found
    patch:
      operationId: updateScanner
      tags: [scanners]
      summary: Update scanner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScannerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scanner"
    delete:
      operationId: deleteScanner
      tags: [scanners]
      summary: Delete scanner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /scanners/{id}/toggle:
    patch:
      operationId: toggleScanner
      tags: [scanners]
      summary: Enable or disable a scanner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScannerToggle"
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scanner"

  /scanners/{id}/scan:
    post:
      operationId: triggerScan
      tags: [scanners]
      summary: Manually trigger a scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"

  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts with optional filters
      parameters:
        - name: scannerId
          in: query
          schema:
            type: integer
        - name: symbol
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"

  /stats/summary:
    get:
      operationId: getStatsSummary
      tags: [stats]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"

  /stats/recent-alerts:
    get:
      operationId: getRecentAlerts
      tags: [stats]
      summary: Get most recent alerts across all scanners
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Recent alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"

  /stats/scanner-activity:
    get:
      operationId: getScannerActivity
      tags: [stats]
      summary: Get per-scanner alert counts over time
      responses:
        "200":
          description: Activity data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScannerActivity"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Scanner:
      type: object
      required: [id, name, chartinkUrl, intervalMinutes, isActive, createdAt, lastScannedAt, alertCount]
      properties:
        id:
          type: integer
        name:
          type: string
        chartinkUrl:
          type: string
        intervalMinutes:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
        lastScannedAt:
          type: ["string", "null"]
        alertCount:
          type: integer
        description:
          type: ["string", "null"]

    ScannerInput:
      type: object
      required: [name, chartinkUrl]
      properties:
        name:
          type: string
          minLength: 1
        chartinkUrl:
          type: string
          minLength: 1
        intervalMinutes:
          type: integer
          default: 5
        description:
          type: string
        isActive:
          type: boolean
          default: true

    ScannerUpdate:
      type: object
      properties:
        name:
          type: string
        chartinkUrl:
          type: string
        intervalMinutes:
          type: integer
        description:
          type: ["string", "null"]
        isActive:
          type: boolean

    ScannerToggle:
      type: object
      required: [isActive]
      properties:
        isActive:
          type: boolean

    Alert:
      type: object
      required: [id, scannerId, scannerName, symbol, triggeredAt]
      properties:
        id:
          type: integer
        scannerId:
          type: integer
        scannerName:
          type: string
        symbol:
          type: string
        price:
          type: ["number", "null"]
        triggeredAt:
          type: string
        telegramSent:
          type: boolean

    ScanResult:
      type: object
      required: [stocksFound, newAlerts, scannerName]
      properties:
        stocksFound:
          type: integer
        newAlerts:
          type: integer
        scannerName:
          type: string

    StatsSummary:
      type: object
      required: [totalScanners, activeScanners, totalAlerts, alertsToday, lastScanAt]
      properties:
        totalScanners:
          type: integer
        activeScanners:
          type: integer
        totalAlerts:
          type: integer
        alertsToday:
          type: integer
        lastScanAt:
          type: ["string", "null"]
        telegramEnabled:
          type: boolean

    ScannerActivity:
      type: object
      required: [scannerId, scannerName, alertCount]
      properties:
        scannerId:
          type: integer
        scannerName:
          type: string
        alertCount:
          type: integer