# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chartink Scanner Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface Alert {
  id: number;
  scannerId: number;
  scannerName: string;
  symbol: string;
  /** @nullable */
  price?: number | null;
  triggeredAt: string;
  telegramSent?: boolean;
}