# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chartink Scanner Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface ScanResult {
  stocksFound: number;
  newAlerts: number;
  scannerName: string;
}