# lib/api-zod/src/generated/types/scanner.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chartink Scanner Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface Scanner {
  id: number;
  name: string;
  chartinkUrl: string;
  intervalMinutes: number;
  isActive: boolean;
  createdAt: string;
  /** @nullable */
  lastScannedAt: string | null;
  alertCount: number;
  /** @nullable */
  description?: string | null;
}