# lib/api-zod/src/generated/types/scannerActivity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chartink Scanner Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface ScannerActivity {
  scannerId: number;
  scannerName: string;
  alertCount: number;
}