# lib/api-zod/src/generated/types/scannerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chartink Scanner Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface ScannerInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  chartinkUrl: string;
  intervalMinutes?: number;
  description?: string;
  isActive?: boolean;
}