# lib/api-zod/src/generated/types/scannerUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chartink Scanner Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface ScannerUpdate {
  name?: string;
  chartinkUrl?: string;
  intervalMinutes?: number;
  /** @nullable */
  description?: string | null;
  isActive?: boolean;
}