# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chartink Scanner Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalScanners: number;
  activeScanners: number;
  totalAlerts: number;
  alertsToday: number;
  /** @nullable */
  lastScanAt: string | null;
  telegramEnabled?: boolean;
}